+++ packages/steps/src/step.tsx
@@ -14,6 +14,8 @@ import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
 import { messages as svMessages } from './locales/sv/messages.mjs';
 
+activateI18n(enMessages, nbMessages, fiMessages, daMessages, svMessages);
+
 const availableAriaLabels = {
   completed: i18n._({
     id: 'steps.aria.completed',
@@ -56,8 +58,6 @@ export interface StepProps {
 }
 
 export function Step(props: StepProps) {
-  activateI18n(enMessages, nbMessages, fiMessages, daMessages, svMessages);
-
   const { active, completed, children } = props;
   const StepsProps = useContext(StepsContext);
   const vertical = !StepsProps.horizontal;